refactor global enums and types
