+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@atoll/shared",
-  "version": "0.41.1-alpha.1",
+  "version": "0.41.1-alpha.2",
   "author": {
     "name": "Kevin Berry",
     "email": "41717340+51ngul4r1ty@users.noreply.github.com"
