+++ src/actions/apiBffViewsBacklogItem.ts
@@ -18,7 +18,7 @@ import type { NoDataApiAction, ApiActionMetaDataRequestMeta } from "../middlewar
 import { buildActionTypes } from "./utils/apiActionUtils";
 
 // interfaces/types
-import { ApiBacklogItem, ApiUserSettings } from "../apiModelTypes";
+import { ApiBacklogItem, ApiUserSettings } from "../types/apiModelTypes";
 
 export interface ApiGetBffViewsBacklogItemResponsePayload {
     response: {
