+++ src/components/organisms/panels/backlogItemPlanning/backlogItemPlanningPanelTypes.ts
@@ -2,8 +2,8 @@
 import { WithTranslation } from "react-i18next";
 
 // interfaces/types
-import { BacklogItemWithSource } from "../../../../reducers/backlogItems/backlogItemsReducerTypes";
-import { StoryPhrases } from "../../../../types";
+import type { BacklogItemWithSource } from "../../../../reducers/backlogItems/backlogItemsReducerTypes";
+import type { StoryPhrases } from "../../../../types/storyTypes";
 
 // consts/enums
 import { BacklogItemType } from "../../../../types/backlogItemTypes";
