+++ src/reducers/userReducer.ts
@@ -7,7 +7,7 @@ import * as ActionTypes from "../actions/actionTypes";
 // interfaces/types
 import { AnyFSA } from "../types/reactHelperTypes";
 import { ActionGetUserPrefsSuccessAction } from "../actions/userActions";
-import { UserSettings } from "../apiModelTypes";
+import { UserSettings } from "../types/apiModelTypes";
 
 export type UserState = Readonly<{
     preferences: UserSettings;
