+++ src/types/dataModelTypes.ts
@@ -1,0 +1,9 @@
+export interface BaseModelItem {
+    id: string;
+}
+
+export interface StandardModelItem extends BaseModelItem {
+    version?: number;
+    createdAt: Date;
+    updatedAt: Date;
+}
