+++ src/types/storyTypes.ts
@@ -1,0 +1,5 @@
+export type StoryPhrases = {
+    rolePhrase: string | null;
+    storyPhrase: string;
+    reasonPhrase: string | null;
+};
