+++ src/utils/wsClient.ts
@@ -1,5 +1,5 @@
 // consts/enums
-import { PushNotificationType } from "../enums";
+import { PushNotificationType } from "../types/pushEnums";
 
 // externals
 import { w3cwebsocket as W3CWebSocket, IMessageEvent } from "websocket";
@@ -8,7 +8,7 @@ import { w3cwebsocket as W3CWebSocket, IMessageEvent } from "websocket";
 import { getApiHostName, getApiPort, requiresSecureProtocol } from "../config";
 
 // interfaces/types
-import type { BasePushNotification } from "../types";
+import type { BasePushNotification } from "../types/pushTypes";
 
 // utils
 import * as logger from "./logger";
